SImplify configs, migrate `vue/*` rules to oxlint

+++ src/create/oxlint.ts
@@ -2,28 +2,34 @@
 
 import fs from 'node:fs/promises';
 import nodePath from 'node:path';
+import { isFileExists } from '../utils.js';
 
-export async function createOxlintConfig(
-	dir: string,
-	// options: { is_node: boolean },
-) {
-	const config = {
-		$schema: './node_modules/oxlint/configuration_schema.json',
-		extends: [
-			'./node_modules/@kirick/lint/configs/oxlint/eslint.jsonc',
-			'./node_modules/@kirick/lint/configs/oxlint/jsdoc.jsonc',
-			'./node_modules/@kirick/lint/configs/oxlint/node.jsonc',
-			'./node_modules/@kirick/lint/configs/oxlint/oxc.jsonc',
-			'./node_modules/@kirick/lint/configs/oxlint/promise.jsonc',
-			'./node_modules/@kirick/lint/configs/oxlint/typescript.jsonc',
-			'./node_modules/@kirick/lint/configs/oxlint/unicorn.jsonc',
-		],
-		ignorePatterns: ['dist'],
-	};
+export async function createOxlintConfig(dir: string) {
+	const path = nodePath.join(dir, 'oxlint.config.ts');
+	if ((await isFileExists(path)) !== true) {
+		await fs.writeFile(
+			nodePath.join(dir, 'oxlint.config.ts'),
+			[
+				`import { rules } from '@kirick/lint/oxlint';`,
+				`import { defineConfig } from 'oxlint';`,
+				'',
+				'export default defineConfig({',
+				'\trules,',
+				`\tignorePatterns: ['dist'],`,
+				'});',
+			].join('\n'),
+			'utf8',
+		);
+	}
 
-	await fs.writeFile(
-		nodePath.join(dir, '.oxlintrc.json'),
-		JSON.stringify(config, null, '\t'),
-		'utf8',
-	);
+	// rename old config
+	{
+		const old_config_path = nodePath.join(dir, '.oxlintrc.json');
+		if (await isFileExists(old_config_path)) {
+			await fs.rename(
+				old_config_path,
+				nodePath.join(dir, '_MIGRATE_OLD_oxlintrc.json'),
+			);
+		}
+	}
 }
